Move avro doc updates from comments to purification
